+++ packages/client-node/src/connection/node_https_connection.ts
@@ -27,7 +27,7 @@ export class NodeHttpsConnection
   protected override buildDefaultHeaders(
     username: string,
     password: string,
-    additional_headers?: Record<string, number | string | string[]>
+    additional_headers?: Record<string, string>
   ): Http.OutgoingHttpHeaders {
     if (this.params.tls?.type === 'Mutual') {
       return {